Refactor StatusColumn into a separate component
